Move MongoDB support classes to their own package
