+++ server/src/test/java/com/ibm/ws/lars/rest/PersistenceBeanLoggingTest.java
@@ -28,6 +28,7 @@ import com.ibm.ws.lars.rest.exceptions.InvalidJsonAssetException;
 import com.ibm.ws.lars.rest.exceptions.NonExistentArtefactException;
 import com.ibm.ws.lars.rest.model.Asset;
 import com.ibm.ws.lars.rest.model.Attachment;
+import com.ibm.ws.lars.rest.mongo.PersistenceBean;
 import com.mongodb.BasicDBObject;
 import com.mongodb.DBCursor;
